updated the candidate formation parameters. changed version to use int, not long

+++ event-aggregation-implementation-parent/event-aggregation-service/src/main/java/atlas/event/aggregation/data/model/ssaevent/Versioned.java
@@ -24,14 +24,14 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public abstract class Versioned
 {
-    private Long version;
+    private int version;
 
-    public Long getVersion()
+    public int getVersion()
     {
         return version;
     }
 
-    public void setVersion(Long version)
+    public void setVersion(int version)
     {
         this.version = version;
     }
